appointment_booking/requirements.txt
google-genai>=1.14.0
telcoflow-sdk>=0.24.0
aiosqlite>=0.20.0
twilio>=9.0.0
python-dotenv>=1.1.0
